requirements.txt
deepgram-sdk
simple-salesforce
